Digital Clock
Task:
For this project, your task is to model a 12-hour and 24-hour digital clock (e.g., 07:34:52 or 22:02:59). The user should be able to set the clock to a specific time, and from there, the clock should accurately begin keeping time. When appropriate, the time should roll over to the next minute, next hour, or both.